WIP following the path to leafreaf that sit anywhere in the tree
In openconfig the convention is that a leafref always points to a "config" container within the parent entry, eg:
        +--rw name             -> ../config/name
        +--rw config
        |  +--rw name?   string
but that is not necessarily the case, as we might have leafrefs to entries in a completely different location in the tree, eg:
        +--rw name             -> ../foo/name
        +--rw name             -> ../../../config/name
        +--rw name             -> /a:conainter1/b:container2/config/name

The changes in "genstate.go" allow to find the referenced entry anywhere it exists in the tree.

Handling of "findMapPaths" in "directory.go" still needs to be handled.